Move the classes to corresponding packages
